requirements.txt
# IRG-Thesis Dependencies

# --- AI APIs ---
google-generativeai>=0.8.0
requests

# --- Vector RAG ---
sentence-transformers>=2.7.0
scikit-learn>=1.3.0

# --- API Server ---
fastapi
uvicorn
python-dotenv

# --- Data & Image Processing ---
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
